[Fix]Work bonus in case of 20000 chars sent!

+++ ft_minitalk.h
@@ -6,7 +6,7 @@
 /*   By: iyamada <iyamada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/17 10:15:54 by iyamada           #+#    #+#             */
-/*   Updated: 2021/12/27 15:50:13 by iyamada          ###   ########.fr       */
+/*   Updated: 2021/12/28 13:22:22 by iyamada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ typedef struct s_receive_info
 	size_t	decimal;
 	bool	is_len_sent;
 	char	*str;
-	int		index;
+	size_t	index;
 }	t_receive_info;
 
 #endif
